refactor: consolidate all exceptions into oopsie/exceptions.py
Move service-layer exceptions from oopsie/services/exceptions.py into
the package-level oopsie/exceptions.py alongside NoInvitationError.
