--- oopsie/services/exceptions.py
@@ -1,25 +1,0 @@
-"""Custom exception hierarchy for Oopsie services."""
-
-
-class OopsieServiceError(Exception):
-    """Base exception for service-layer errors."""
-
-
-class GitOperationError(OopsieServiceError):
-    """Raised when a git CLI operation fails."""
-
-
-class ClaudeCodeError(OopsieServiceError):
-    """Raised when Claude Code CLI fails or times out."""
-
-
-class GitHubApiError(OopsieServiceError):
-    """Raised when GitHub REST API call fails."""
-
-
-class GitHubAppNotConfiguredError(OopsieServiceError):
-    """Raised when GitHub App credentials are not configured."""
-
-
-class AnthropicKeyNotConfiguredError(OopsieServiceError):
-    """Raised when no Anthropic API key is configured for a project or its org."""
