+++ tests/web/test_projects.py
@@ -4,12 +4,12 @@ import uuid
 from unittest.mock import AsyncMock, patch
 
 import pytest
+from oopsie.exceptions import GitHubApiError
 from oopsie.models import Project
 from oopsie.services.anthropic_key_service import (
     get_anthropic_api_key,
     set_anthropic_api_key,
 )
-from oopsie.services.exceptions import GitHubApiError
 from sqlalchemy import select
 from sqlalchemy.ext.asyncio import AsyncSession
